separeta TFFMBaseModel class
